+++ app/__init__.py
@@ -12,6 +12,6 @@ from .config_loaders.mongodb_client_configured import mongodb_client_configured
 from .config_loaders.tiktok_cookie_configured import tiktok_cookie_configured
 from .process_resource import process_resource
 from .models.resource_audio import ResourceAudio
-from .resource_message_to_json import resource_message_to_json
-from .resource_json_to_input_item import resource_json_to_input_item
+from .converters.resource_message_to_json import resource_message_to_json
+from .converters.resource_json_to_input_item import resource_json_to_input_item
 from .models.subtitle import Subtitle
