+++ app/converters/resource_json_to_input_item.py
@@ -1,8 +1,8 @@
 from app.input_items.input_item import InputItem
-from .input_items.input_item import InputItem
-from .input_items.input_item_hosted import InputItemHosted
-from .input_items.input_item_tiktok import InputItemTiktok
-from .input_items.input_item_youtube import InputItemYoutube
+from ..input_items.input_item import InputItem
+from ..input_items.input_item_hosted import InputItemHosted
+from ..input_items.input_item_tiktok import InputItemTiktok
+from ..input_items.input_item_youtube import InputItemYoutube
 
 INPUT_TYPE_TO_ITEM_CLASS = {"hosted": InputItemHosted,
                             "tiktok": InputItemTiktok,
